B_Members now have transform
Though we have the correct place we need to update these trasnforms in their construction we still need to find a way to either bring the entities we created in game.ccp to our skeletal hierarchy. I think this should be done by passing in a dictionary of type <string, std...Transform> which can then be organized inside.